refactoring to move the swap_bundle.rs file into the bundle.rs file
